C2_W3_Assignment_updated_2023
So, in the latest version of the labs, we cannot add code in the cell except for the cells marked with ###START CODE HERE  and ###END CODE HERE.
In your code/Notebook, you have imported the required libraries in the import cell(1st cell). However, that won't be possible in the latest version. So, I have added code to import libraries at the appropriate places where it is used.